haari/requirements.txt
django
djangorestframework
django-tailwind
#'django-tailwind[reload]'
mysqlclient
